+++ app/models/noticed/ephemeral.rb
@@ -59,3 +59,5 @@ module Noticed
     end
   end
 end
+
+ActiveSupport.run_load_hooks :noticed_ephemeral, Noticed::Ephemeral