+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 ### Unreleased
 
+### 2.6.1
+
+* Handle text/html and application/json responses from Slack #531
+* Handle 400 errors for FCM deliveries #532
+* Add lazy load hooks for models
+
 ### 2.6.0
 
 * Add `mark_as_read_and_seen` and `mark_as_unread_and_unseen` class methods to `Noticed::Notification`